requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
gTTS==2.2.3
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.7.1
newspaper3k==0.2.8
nltk==3.6.7
Pillow==9.0.0
pipreqs==0.4.11
playsound==1.3.0
python-dateutil==2.8.2
python-vlc==3.0.12118
PyYAML==6.0
regex==2021.11.10
requests==2.27.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
urllib3==1.26.7
yarg==0.1.9
